requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
psycopg2==2.9.9
python-dotenv==1.0.1
shortuuid==1.0.11
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
